move more utility functions to utils.go
